pyproject.toml
[project]
name = "orchestra-backend"
version = "0.1.0"
description = "Orchestra multi-agent AI platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "litellm>=1.40.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.10.1",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "asyncpg>=0.30.0",
    "motor>=3.6.0",
    "neo4j>=5.25.0",
    "hvac>=2.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "redis[hiredis]>=5.0.0",
    "google-adk>=1.28.1",
    "requests>=2.33.1",
    "python-gitlab>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
